Local build works with hard coded branch name(#5994)
