+++ packages/utilities/fast-benchmarks/.gitignore
@@ -1,4 +1,3 @@
 tsconfig.tsbuildinfo
 node_modules
 dist
-yarn.lock
