--- packages/utilities/fast-benchmarks/results/result.json
@@ -1,148 +1,0 @@
-{
-  "benchmarks": [
-    {
-      "name": "my-benchmark-createDelete5x-memory",
-      "bytesSent": 2462,
-      "version": "1.8.0",
-      "measurement": {
-        "name": "usedJSHeapSize",
-        "mode": "expression",
-        "expression": "window.usedJSHeapSize"
-      },
-      "browser": {
-        "name": "chrome",
-        "headless": true,
-        "windowSize": {
-          "width": 1024,
-          "height": 768
-        },
-        "addArguments": [
-          "--js-flags=--expose-gc",
-          "--enable-precise-memory-info"
-        ],
-        "userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/101.0.4951.64 Safari/537.36"
-      },
-      "mean": {
-        "low": 2.798855688350844,
-        "high": 2.8072279116491567
-      },
-      "differences": [
-        null,
-        {
-          "absolute": {
-            "low": -0.006109477419492123,
-            "high": 0.006640144086160134
-          },
-          "percentChange": {
-            "low": -0.21799172212870524,
-            "high": 0.2369253286071103
-          }
-        }
-      ],
-      "samples": [
-        2.808279,
-        2.801631,
-        2.802211,
-        2.808047,
-        2.810679,
-        2.808827,
-        2.804287,
-        2.812383,
-        2.781763,
-        2.810795,
-        2.789987,
-        2.794883,
-        2.811811,
-        2.783375,
-        2.777799,
-        2.806811,
-        2.808839,
-        2.809983,
-        2.812307,
-        2.808351,
-        2.779391,
-        2.808427,
-        2.812959,
-        2.810375,
-        2.811219,
-        2.811399,
-        2.797831,
-        2.809859,
-        2.811671,
-        2.785075
-      ]
-    },
-    {
-      "name": "my-benchmark-createDelete5x-memory",
-      "bytesSent": 2462,
-      "version": "1.9.0",
-      "measurement": {
-        "name": "usedJSHeapSize",
-        "mode": "expression",
-        "expression": "window.usedJSHeapSize"
-      },
-      "browser": {
-        "name": "chrome",
-        "headless": true,
-        "windowSize": {
-          "width": 1024,
-          "height": 768
-        },
-        "addArguments": [
-          "--js-flags=--expose-gc",
-          "--enable-precise-memory-info"
-        ],
-        "userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/101.0.4951.64 Safari/537.36"
-      },
-      "mean": {
-        "low": 2.7979686936484023,
-        "high": 2.8075842396849304
-      },
-      "differences": [
-        {
-          "absolute": {
-            "low": -0.006640144086160134,
-            "high": 0.006109477419492123
-          },
-          "percentChange": {
-            "low": -0.23688137250091656,
-            "high": 0.2179495582601153
-          }
-        },
-        null
-      ],
-      "samples": [
-        2.809623,
-        2.810455,
-        2.811059,
-        2.810767,
-        2.805663,
-        2.775991,
-        2.811427,
-        2.812123,
-        2.778291,
-        2.812443,
-        2.792823,
-        2.775027,
-        2.811891,
-        2.811607,
-        2.810131,
-        2.810563,
-        2.809639,
-        2.809727,
-        2.811411,
-        2.805347,
-        2.810707,
-        2.805539,
-        2.785347,
-        2.783935,
-        2.776715,
-        2.811763,
-        2.812499,
-        2.807943,
-        2.796287,
-        2.806551
-      ]
-    }
-  ]
-}
