+++ packages/utilities/fast-benchmarks/scripts/template.js
@@ -246,27 +246,9 @@ async function generateBenchmarks(
                 };
             }
 
-            //adjust some settings to separately report memory benchmark results
-            // const memoryBench = JSON.parse(JSON.stringify(bench));
-            // const memoryMeasurement = [
-            //     {
-            //         name: "usedJSHeapSize",
-            //         mode: "expression",
-            //         expression: "window.usedJSHeapSize",
-            //     },
-            // ];
-            // memoryBench.name = `${name}-memory`;
-            // memoryBench.measurement = memoryMeasurement;
-            // memoryBench.browser.addArguments = [
-            //     "--js-flags=--expose-gc",
-            //     "--enable-precise-memory-info",
-            // ];
-            // memoryBenchmarks.push(memoryBench);
-
             if (!memory) benchmarks.push(bench);
         });
 
-        // tachoData[`${operation}-memory`] = memoryBenchmarks;
         if (!memory) tachoData[operation] = benchmarks;
     });
     return tachoData;
@@ -327,11 +309,11 @@ export async function generateTemplates(options) {
     try {
         const tsConfigPath = await generateTsConfig(options);
         const fileName = `${options.library}_${options.benchmark}`;
-
         //special handling if 'local' version was passed in as an option
         const localProps = { branchName: "", operationProps: {} };
         if (options.versions.includes(LOCAL)) {
-            localProps.branchName = await getLocalGitBranchName();
+            // localProps.branchName = await getLocalGitBranchName();
+            localProps.branchName = "users/wendy/local-build-actions";
             // check if user passed in localBenchFile for different implementation of local
             if (options.localBenchFile)
                 localProps.operationProps = await generateHtmlTemplates(
